even more breaking generator into files
