+++ src/stan/lang/generator/generate_public_decl.hpp
@@ -12,7 +12,7 @@ namespace stan {
      * Generate the public declaration scope for a class to the
      * specified stream.
      *
-     * @param[in,out] stream for generating
+     * @param[in,out] o stream for generating
      */
     void generate_public_decl(std::ostream& o) {
       o << "public:" << EOL;
